% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenwood-package.R
\docType{package}
\name{greenwood-package}
\alias{greenwood}
\alias{greenwood-package}
\title{greenwood: Modern Survival Analysis}
\description{
Greenwood is developed as an R and Python monorepo. Both packages share one
response model and one set of contracts:
}
\details{
\itemize{
\item \code{\link[=Surv]{Surv()}} is \code{survival::Surv()}, re-exported unchanged.
}

The Python package implements the same function under the same name. The
contracts and the cases both test suites replay live in the repository's
\verb{spec/} directory.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rich-iannone/greenwood}
  \item Report bugs at \url{https://github.com/rich-iannone/greenwood/issues}
}

}
\author{
\strong{Maintainer}: Richard Iannone \email{riannone@me.com} (\href{https://orcid.org/0000-0003-3925-190X}{ORCID})

}
\keyword{internal}
