% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenwood-package.R
\docType{package}
\name{greenwood-package}
\alias{greenwood}
\alias{greenwood-package}
\title{greenwood: Modern Survival Analysis}
\description{
Greenwood provides tools for survival analysis, the study of time-to-event
data. \code{\link[=Surv]{Surv()}} builds the survival response that models and estimators take
as their outcome.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rich-iannone/greenwood}
  \item Report bugs at \url{https://github.com/rich-iannone/greenwood/issues}
}

}
\author{
\strong{Maintainer}: Richard Iannone \email{riannone@me.com} (\href{https://orcid.org/0000-0003-3925-190X}{ORCID})

}
\keyword{internal}
